+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: tmazitov <tmazitov@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/03 13:23:25 by tmazitov          #+#    #+#             */
-/*   Updated: 2023/07/06 21:28:00 by tmazitov         ###   ########.fr       */
+/*   Updated: 2023/07/14 14:59:31 by tmazitov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
-	int	counter;
+	size_t	counter;
 
 	counter = 0;
 	while (str[counter])
